tests/requirements.txt
pytest>=8.0.0
pytest-cov>=4.1.0
httpx>=0.25.1
pandas>=2.1.1
numpy>=1.24.3
scikit-learn>=1.3.0
joblib>=1.3.2
PyYAML>=6.0.1
fastapi>=0.109.0
uvicorn>=0.27.0 